+++ src/features/repository/RepositoryView.tsx
@@ -127,7 +127,7 @@ export function RepositoryView() {
   const activeSecondary = SECONDARY_TABS.find((t) => t.tab === repoTab);
 
   return (
-    <div className="flex h-screen flex-col">
+    <div className="flex min-h-0 flex-1 flex-col">
       <RepoHeader repoPath={repoPath} />
       <div className="flex min-h-0 flex-1">
         <aside className="flex w-96 shrink-0 flex-col border-r">